rvvm: Make pending_irqs 32-bit
- AIA doesn't actually introduce new external interrupt in mip/sip CSRs
- Everything else is hart local anyways

+++ src/rvvm.h
@@ -253,7 +253,7 @@ struct rvvm_hart_t {
     rvtimecmp_t mtimecmp;
     rvtimecmp_t stimecmp;
 
-    uint64_t pending_irqs;
+    uint32_t pending_irqs;
     uint32_t pending_events;
     uint32_t preempt_ms;
